@@ -1,8 +1,8 @@
-package com.anyhunter.app.location;
+package com.hypeagle.app.location;
 
 import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.platform.app.InstrumentationRegistry;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -21,6 +21,6 @@ public class ExampleInstrumentedTest {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getTargetContext();
 
-        assertEquals("com.anyhunter.app.location.test", appContext.getPackageName());
+        assertEquals("com.hypeagle.app.location.test", appContext.getPackageName());
     }
 }
